% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{apply_quantile_normalisation}
\alias{apply_quantile_normalisation}
\title{Internal Helper: Apply Quantile Normalisation}
\usage{
apply_quantile_normalisation(binding_profiles_data, quantile_norm)
}
\arguments{
\item{binding_profiles_data}{The data.frame of binding profiles.}

\item{quantile_norm}{Logical. If TRUE, normalisation is applied.}
}
\value{
A data.frame, which is quantile-normalised if requested.
}
\description{
This internal function applies quantile normalisation to a binding profile
data.frame if the user requests it. It is not exported.
}
\keyword{internal}
