% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{gene_occupancy}
\alias{gene_occupancy}
\title{Compute average occupancy for each gene, using DamID binding profiles}
\usage{
gene_occupancy(
  binding_data,
  ensdb_genes = get_ensdb_genes()$genes,
  buffer = 0,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{binding_data}{A data.frame as returned by build_dataframes(): Must have columns 'chr', 'start', 'end', and then sample columns.}

\item{ensdb_genes}{A GRanges object of gene annotations.  Retrieved by `get_ensdb_genes` by default, but a custom GRanges object of genes can be provided.  Must contain metadata columns 'gene_name' and 'gene_id'.}

\item{buffer}{Integer. Optional number of bases to extend gene boundaries both upstream and downstream.}

\item{BPPARAM}{A BiocParallel parameter object. Default: BiocParallel::bpparam(). On Windows, will use serial for safety.}
}
\value{
data.frame: rows = genes, columns = one per sample, plus annotation columns (gene name, nfrags, etc.)
}
\description{
For each gene, averages log2 scores of all overlapping fragments (weighted by coverage).
}
