% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{import_bedgraph_as_df}
\alias{import_bedgraph_as_df}
\title{Import a bedGraph file as a data.frame (chr, start, end, value)}
\usage{
import_bedgraph_as_df(path, colname = "score")
}
\arguments{
\item{path}{File path (bedGraph)}

\item{colname}{Name of the value column (usually sample name).}
}
\value{
data.frame
}
\description{
Import a bedGraph file as a data.frame (chr, start, end, value)
}
\keyword{internal}
