% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{import_peaks}
\alias{import_peaks}
\title{Import a GFF file as a GRanges object}
\usage{
import_peaks(path)
}
\arguments{
\item{path}{File path (GFF/GTF)}
}
\value{
GRanges
}
\description{
Import a GFF file as a GRanges object
}
\keyword{internal}
