% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{locate_files}
\alias{locate_files}
\title{Locate files with optional wildcard expansion}
\usage{
locate_files(paths, pattern = NULL)
}
\arguments{
\item{paths}{Character vector of directories or patterns (e.g., "/mypath/sample*").}

\item{pattern}{Optional regex for file extensions (e.g., "\\.gff$").  Case is ignored.}
}
\value{
Character vector of file paths.
}
\description{
Locate files with optional wildcard expansion
}
\keyword{internal}
