% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-DamIDResults.R
\name{metadata,DamIDResults-method}
\alias{metadata,DamIDResults-method}
\title{Access analysis metadata}
\usage{
\S4method{metadata}{DamIDResults}(x, ...)
}
\arguments{
\item{x}{A \code{DamIDResults} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A nested \code{list} of metadata.
}
\description{
This method returns the provenance and annotation metadata for the
  analysis. It overrides the default \code{metadata} method for
  \code{DamIDResults} objects.
}
