% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{plot_catada_mean_variance}
\alias{plot_catada_mean_variance}
\title{Plot mean-variance relationship for CATaDa results}
\usage{
plot_catada_mean_variance(diff_results)
}
\arguments{
\item{diff_results}{A DamIDResults object.}
}
\value{
A ggplot object.
}
\description{
This diagnostic function assesses whether CATaDa occupancy signal follows
the theoretical mean-variance relationship assumed by Negative Binomial
models. It is only applicable to objects generated via the
`differential_accessibility()` function.
}
