% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Draw Proportional Venn Diagram for differential binding analysis}
\usage{
plot_venn(
  diff_results,
  title = "Enriched binding at loci",
  subtitle = "",
  set_labels = NULL,
  filename = NULL,
  font = "sans",
  format = c("pdf", "svg"),
  region_colours = c("#FFA500", "#2288DD", "#CCCCCC")
)
}
\arguments{
\item{diff_results}{A `DamIDResults` object, as returned by
`differential_binding()` or `differential_accessibility()`.}

\item{title}{Plot title to use.}

\item{subtitle}{Subtitle to use (default is empty).}

\item{set_labels}{Character vector of length 2. Names for the two sets/circles (defaults to the analysis condition names).}

\item{filename}{Character. Path at which to save the diagram, if not NULL.}

\item{font}{Font name to use (default is "sans")}

\item{format}{Character. Output plot format, "pdf" or "svg" (default "pdf").}

\item{region_colours}{Character vector of length 2 or 3. Fill colours for each set region (default: c("#FFA500", "#2288DD", "#CCCCCC")).}
}
\value{
Invisibly, the output of `BioVenn::draw.venn()`.
}
\description{
Generates a two-set Venn/proportional diagram summarising the results of the differential binding analysis.
The set union represents significant binding peaks that fail to show significant differences in occupancy;
the exclusive regions of each set represent regions with enriched differential binding in that condition.
Note that regions can be bound in both conditions, and still show differential occupancy.
}
