% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataframes.R
\name{process_binding_profiles}
\alias{process_binding_profiles}
\title{Internal Helper: Process Binding Profile Inputs}
\usage{
process_binding_profiles(binding_profiles_path = NULL, binding_profiles = NULL)
}
\arguments{
\item{binding_profiles_path}{Character vector. Path(s) to directories or file globs.}

\item{binding_profiles}{List of GRanges objects.}
}
\value{
A dataframe containing the merged binding profiles.
}
\description{
This internal function encapsulates the logic for loading binding profiles,
whether from file paths or from a list of GRanges objects. It performs
validation and returns a unified data.frame.
}
\keyword{internal}
