% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_functions.R
\name{reduce_regions}
\alias{reduce_regions}
\title{Reduce a list of GRanges to unique, non-overlapping regions}
\usage{
reduce_regions(peaks)
}
\arguments{
\item{peaks}{List of GRanges objects (e.g., from multiple samples)}
}
\value{
GRanges: Reduced (union) regions.
}
\description{
Reduce a list of GRanges to unique, non-overlapping regions
}
